+++ src/aceinna/devices/ping/ping_tool.py
@@ -1,4 +1,4 @@
-from .rtk330l import ping as ping_opendevice
+from .rtk330l import ping as ping_rtk330l
 from .ins401 import ping as ping_ins401
 from ...framework.context import APP_CONTEXT
 from ...framework.constants import INTERFACES
@@ -8,7 +8,7 @@ def do_ping(communicator_type, device_access, filter_device_type):
         if filter_device_type is None or filter_device_type in ['RTKL']:
             APP_CONTEXT.get_logger().logger.debug(
                 'Checking if is RTK330L device...')
-            ping_result = ping_opendevice(
+            ping_result = ping_rtk330l(
                 device_access, filter_device_type)
             if ping_result:
                 return ping_result