+++ HISTORY.md
@@ -12,5 +12,5 @@
 ## 2.6.0，08/19/2021
 
 - [INS401] Add the Ethernet driver for INS401 device, supports data logging on a PC using a 100base-TX to 100base-T1 converter
-- [INS401,RTK330LA, OpenRTK330LI] New code developments based on OpenRTK Python driver 2.5.0
+- [INS401,RTK330LA] New code developments based on OpenRTK Python driver 2.5.0